docs(context-manager): split into overview and custom strategies pages
Split context-management.mdx into two pages:
- Overview (index.mdx): auto, agentic, disabling, limitations
- Custom Strategies (custom-strategies.mdx): ContextManager instance,
  Offload builder, targets, conditions, stash config

Update navigation.yml for the new directory structure and fix
cross-reference links in deprecation callouts.
