+++ site/src/content/docs/user-guide/concepts/agents/conversation-management.mdx
@@ -36,7 +36,7 @@ See [Context Management](../context-management).
 
 :::caution[Deprecated]
 `ConversationManager` is deprecated and will be removed in
-v2. Use [ContextManager](../context-management#custom-strategies)
+v2. Use [ContextManager](../context-management/custom-strategies)
 instead for strategy-driven context management, or pass
 <Syntax py='context_manager="auto"' ts='contextManager: "auto"' />
 for managed defaults.
