+++ site/src/content/docs/user-guide/concepts/context-management/custom-strategies.ts
@@ -2,39 +2,8 @@ import {
   Agent,
   ContextManager,
   Offload,
-  SlidingWindowConversationManager,
 } from '@strands-agents/sdk'
 import { LocalFileStorage } from '@strands-agents/sdk/storage'
-
-async function basic() {
-  // --8<-- [start:basic]
-  const agent = new Agent({
-    contextManager: 'auto',
-  })
-  // --8<-- [end:basic]
-}
-
-async function agentic() {
-  // --8<-- [start:agentic]
-  const agent = new Agent({
-    contextManager: 'agentic',
-  })
-  // --8<-- [end:agentic]
-}
-
-async function customConversationManager() {
-  // --8<-- [start:custom_conversation_manager]
-  // Your conversation manager is used;
-  // ContextOffloader is still added automatically
-  const agent = new Agent({
-    contextManager: 'auto',
-    conversationManager:
-      new SlidingWindowConversationManager({
-        windowSize: 30,
-      }),
-  })
-  // --8<-- [end:custom_conversation_manager]
-}
 
 async function customStrategies() {
   // --8<-- [start:custom_strategies]
@@ -70,12 +39,4 @@ async function stash() {
     }),
   })
   // --8<-- [end:stash]
-}
-
-async function disabled() {
-  // --8<-- [start:disabled]
-  const agent = new Agent({
-    contextManager: false,
-  })
-  // --8<-- [end:disabled]
 }
