+++ site/src/content/docs/user-guide/concepts/context-management/custom-strategies_imports.ts
@@ -1,0 +1,20 @@
+// @ts-nocheck
+
+// --8<-- [start:custom_strategies_imports]
+import {
+  Agent,
+  ContextManager,
+  Offload,
+} from '@strands-agents/sdk'
+// --8<-- [end:custom_strategies_imports]
+
+// --8<-- [start:stash_imports]
+import {
+  Agent,
+  ContextManager,
+  Offload,
+} from '@strands-agents/sdk'
+import {
+  LocalFileStorage,
+} from '@strands-agents/sdk/storage'
+// --8<-- [end:stash_imports]
