+++ site/src/content/docs/user-guide/concepts/context-management/index.mdx
@@ -6,7 +6,6 @@ sourceLinks:
   - path: strands-py/src/strands/_context_manager/modes/agentic/agentic_context.py
   - path: strands-py/src/strands/agent/conversation_manager/conversation_manager.py
   - path: strands-ts/src/context-manager/context-manager.ts
-  - path: strands-ts/src/context-manager/strategies/offload/index.ts
   - path: strands-ts/src/conversation-manager/conversation-manager.ts
 addedDate: 2026-06-12
 ---
@@ -19,8 +18,9 @@ performance, and loss of relevant information.
 The SDK ships context management configurations that work
 out of the box, so you don't have to assemble conversation
 managers, offloaders, and thresholds by hand. Pick a mode
-and the SDK wires up the pieces with tuned defaults. You can
-still drop down to explicit configuration when you need it.
+and the SDK wires up the pieces with tuned defaults. For
+full control, see
+[Custom Strategies](./custom-strategies).
 
 ## Automatic context management
 
@@ -44,9 +44,9 @@ agent = Agent(context_manager="auto")
 <Tab label="TypeScript">
 
 ```typescript
---8<-- "user-guide/concepts/context-management_imports.ts:basic_imports"
+--8<-- "user-guide/concepts/context-management/index_imports.ts:basic_imports"
 
---8<-- "user-guide/concepts/context-management.ts:basic"
+--8<-- "user-guide/concepts/context-management/index.ts:basic"
 ```
 </Tab>
 </Tabs>
@@ -69,8 +69,8 @@ evaluations relative to other Strands Agent configurations:
   full content on demand.
 
 For full details on each component, see
-[Conversation Management](./agents/conversation-management#summarizingconversationmanager)
-and [Context Offloader](./plugins/context-offloader).
+[Conversation Management](../agents/conversation-management#summarizingconversationmanager)
+and [Context Offloader](../plugins/context-offloader).
 
 ### Combining with explicit configuration
 
@@ -110,9 +110,9 @@ agent = Agent(
 <Tab label="TypeScript">
 
 ```typescript
---8<-- "user-guide/concepts/context-management_imports.ts:custom_conversation_manager_imports"
+--8<-- "user-guide/concepts/context-management/index_imports.ts:custom_conversation_manager_imports"
 
---8<-- "user-guide/concepts/context-management.ts:custom_conversation_manager"
+--8<-- "user-guide/concepts/context-management/index.ts:custom_conversation_manager"
 ```
 </Tab>
 </Tabs>
@@ -146,9 +146,9 @@ agent = Agent(context_manager="agentic")
 <Tab label="TypeScript">
 
 ```typescript
---8<-- "user-guide/concepts/context-management_imports.ts:agentic_imports"
+--8<-- "user-guide/concepts/context-management/index_imports.ts:agentic_imports"
 
---8<-- "user-guide/concepts/context-management.ts:agentic"
+--8<-- "user-guide/concepts/context-management/index.ts:agentic"
 ```
 </Tab>
 </Tabs>
@@ -221,87 +221,6 @@ message rather than compressing on a fixed threshold. The
 tradeoff is the tokens the model spends reading telemetry
 and calling the tools.
 
-## Custom strategies
-
-:::caution[Experimental]
-`ContextManager` is an experimental API. The strategy
-interface and builder methods may change in future versions.
-:::
-
-:::note
-Custom `ContextManager` instances are currently available in
-the TypeScript SDK only. Python support is planned.
-:::
-
-When the built-in modes don't fit, pass a `ContextManager`
-instance directly. You define the strategy pipeline: what
-gets offloaded, when, and how.
-
-```typescript
---8<-- "user-guide/concepts/context-management_imports.ts:custom_strategies_imports"
-
---8<-- "user-guide/concepts/context-management.ts:custom_strategies"
-```
-
-Strategies run as an ordered pipeline. Each strategy sees
-the output of the previous one, so order matters: if an
-earlier strategy shrinks content below the next strategy's
-threshold, the later strategy skips it automatically.
-
-### The Offload builder
-
-`Offload` provides three methods, each targeting different
-content in the context window:
-
-| Method | What it does |
-|--------|-------------|
-| `Offload.truncate(target)` | Replaces content with a head/tail preview |
-| `Offload.summarize(target)` | Replaces content with a model-generated summary |
-| `Offload.drop(target)` | Removes content entirely |
-
-**Targets** control what content the strategy acts on:
-
-- `"toolResults"`: all successful tool result blocks
-- `"toolResultErrors"`: failed tool result blocks
-- `"assistantText"`: text blocks in assistant messages
-- `"*"`: all content in the context window
-- `string[]`: tool results from specific tools
-  (e.g. `["tool::bash", "tool::read_file"]`)
-
-Chain `.when()` to set conditions:
-
-- `threshold`: per-block token threshold (strategy fires
-  on each block above this size)
-- `utilization`: context window ratio (0 to 1+) above
-  which the strategy fires
-- `preserveRecent`: number of recent matching messages
-  to leave untouched
-
-Setting `threshold` alone makes the strategy act eagerly,
-on each message as it arrives. Setting `utilization` alone
-makes it act in batch when the window fills. Both together
-target oversized blocks, but only when the window is
-pressured.
-
-### Stash configuration
-
-When a strategy offloads content, it stores it in the
-stash so the agent can retrieve it later via a
-`retrieve_context` tool. By default, the stash uses
-in-memory storage. To persist across restarts, provide
-a durable storage backend:
-
-```typescript
---8<-- "user-guide/concepts/context-management_imports.ts:stash_imports"
-
---8<-- "user-guide/concepts/context-management.ts:stash"
-```
-
-Set `stash: false` to disable the stash entirely (no
-persistence, no retrieval tool). Set
-`stash: { retrievalTool: false }` to keep persistence but
-hide the retrieval tool from the agent.
-
 ## Disabling context management
 
 To turn off all context management (no compression, no
@@ -319,9 +238,9 @@ agent = Agent(context_manager=False)
 <Tab label="TypeScript">
 
 ```typescript
---8<-- "user-guide/concepts/context-management_imports.ts:disabled_imports"
+--8<-- "user-guide/concepts/context-management/index_imports.ts:disabled_imports"
 
---8<-- "user-guide/concepts/context-management.ts:disabled"
+--8<-- "user-guide/concepts/context-management/index.ts:disabled"
 ```
 </Tab>
 </Tabs>
@@ -341,7 +260,8 @@ mode with a stateful model raises a
 The `"auto"` and `"agentic"` modes compose an offloader
 that uses in-memory storage, which does not persist across
 process restarts. If your agent uses session management and
-needs durable offloaded content, configure a custom
-`ContextManager` with a durable stash (TypeScript), or an
-explicit `ContextOffloader` with file or S3 storage. See
-[Storage Backends](./plugins/context-offloader#storage-backends).
+needs durable offloaded content, configure a
+[custom ContextManager with a durable stash](./custom-strategies#stash-configuration)
+(TypeScript), or an explicit `ContextOffloader` with file
+or S3 storage. See
+[Storage Backends](../plugins/context-offloader#storage-backends).
