+++ site/src/content/docs/user-guide/concepts/context-management/index_imports.ts
@@ -15,25 +15,6 @@ import {
 } from '@strands-agents/sdk'
 // --8<-- [end:custom_conversation_manager_imports]
 
-// --8<-- [start:custom_strategies_imports]
-import {
-  Agent,
-  ContextManager,
-  Offload,
-} from '@strands-agents/sdk'
-// --8<-- [end:custom_strategies_imports]
-
-// --8<-- [start:stash_imports]
-import {
-  Agent,
-  ContextManager,
-  Offload,
-} from '@strands-agents/sdk'
-import {
-  LocalFileStorage,
-} from '@strands-agents/sdk/storage'
-// --8<-- [end:stash_imports]
-
 // --8<-- [start:disabled_imports]
 import { Agent } from '@strands-agents/sdk'
 // --8<-- [end:disabled_imports]
