+++ site/src/content/docs/user-guide/concepts/plugins/context-offloader.mdx
@@ -10,7 +10,7 @@ addedDate: 2026-04-24
 
 :::caution[Deprecated]
 `ContextOffloader` is deprecated and will be removed in v2.
-Use [ContextManager](../context-management#custom-strategies)
+Use [ContextManager](../context-management/custom-strategies)
 with `Offload` strategies instead, or pass
 <Syntax py='context_manager="auto"' ts='contextManager: "auto"' />
 for managed defaults.
